feat: extend sink plugin with info
