+++ agent/agent_test.go
@@ -79,7 +79,7 @@ func TestRunnerRun(t *testing.T) {
 
 		mSink := new(mockPassthroughSink)
 		sinkFactory := registry.NewSinkFactory()
-		sinkFactory.Register("mock-sink", newMockSinkFn(mSink), "")
+		sinkFactory.Register("mock-sink", newMockSinkFn(mSink))
 
 		r := agent.NewAgent(extrFactory, procFactory, sinkFactory, nil, test.Logger)
 		run := r.Run(rcp)
@@ -94,7 +94,7 @@ func TestRunnerRun(t *testing.T) {
 
 		mSink := new(mockPassthroughSink)
 		sinkFactory := registry.NewSinkFactory()
-		sinkFactory.Register("mock-sink", newMockSinkFn(mSink), "")
+		sinkFactory.Register("mock-sink", newMockSinkFn(mSink))
 
 		r := agent.NewAgent(extrFactory, procFactory, sinkFactory, nil, test.Logger)
 		run := r.Run(rcp)
@@ -104,7 +104,7 @@ func TestRunnerRun(t *testing.T) {
 	t.Run("should run extractor, processors and sinks", func(t *testing.T) {
 		mSink := new(mockPassthroughSink)
 		sinkFactory := registry.NewSinkFactory()
-		sinkFactory.Register("mock-sink", newMockSinkFn(mSink), "")
+		sinkFactory.Register("mock-sink", newMockSinkFn(mSink))
 
 		r := agent.NewAgent(extrFactory, procFactory, sinkFactory, nil, test.Logger)
 		run := r.Run(validRecipe)
@@ -117,7 +117,7 @@ func TestRunnerRun(t *testing.T) {
 	t.Run("should record metrics", func(t *testing.T) {
 		mSink := new(mockPassthroughSink)
 		sinkFactory := registry.NewSinkFactory()
-		sinkFactory.Register("mock-sink", newMockSinkFn(mSink), "")
+		sinkFactory.Register("mock-sink", newMockSinkFn(mSink))
 		monitor := new(mockMonitor)
 		monitor.On("RecordRun", validRecipe, mock.AnythingOfType("int"), true).Once()
 		defer monitor.AssertExpectations(t)
@@ -142,7 +142,7 @@ func TestRunnerRunMultiple(t *testing.T) {
 		recipeList := []recipe.Recipe{validRecipe, failedRecipe, validRecipe2}
 
 		sinkFactory := registry.NewSinkFactory()
-		sinkFactory.Register("mock-sink", newMockSinkFn(new(mockPassthroughSink)), "")
+		sinkFactory.Register("mock-sink", newMockSinkFn(new(mockPassthroughSink)))
 
 		r := agent.NewAgent(extrFactory, procFactory, sinkFactory, nil, test.Logger)
 		runs := r.RunMultiple(recipeList)
@@ -167,8 +167,8 @@ func TestRunnerRunMultiple(t *testing.T) {
 		sink1 := new(mockPassthroughSink)
 		sink2 := new(mockPassthroughSink)
 		sinkFactory := registry.NewSinkFactory()
-		sinkFactory.Register("mock-sink", newMockSinkFn(sink1), "")
-		sinkFactory.Register("mock-sink-2", newMockSinkFn(sink2), "")
+		sinkFactory.Register("mock-sink", newMockSinkFn(sink1))
+		sinkFactory.Register("mock-sink-2", newMockSinkFn(sink2))
 
 		r := agent.NewAgent(extrFactory, procFactory, sinkFactory, nil, test.Logger)
 		r.RunMultiple(recipeList)
@@ -237,7 +237,11 @@ func newMockSinkFn(sink plugins.Syncer) func() plugins.Syncer {
 	}
 }
 
-func (m *mockPassthroughSink) ValidateConfig(config map[string]interface{}) error {
+func (m *mockPassthroughSink) Info() (plugins.Info, error) {
+	return plugins.Info{}, nil
+}
+
+func (m *mockPassthroughSink) Validate(config map[string]interface{}) error {
 	return nil
 }
