+++ plugins/extractors/clickhouse/meta.yaml
@@ -1,6 +1,6 @@
-name: "clickhouse"
-version: "1.0.0"
-description: "descr"
+name: "ClickHouse"
+version: "v1"
+description: "Column-oriented DBMS for online analytical processing."
 sample_config: |-
   host: localhost:9000
   user_id: admin
