+++ plugins/extractors/csv/meta.yaml
@@ -1,5 +1,5 @@
-name: "csv"
-version: "1.0.0"
-description: "descr"
+title: "csv"
+version: "v1"
+description: "Comma-separated values file."
 sample_config: |-
   path: ./path-to-a-file-or-a-directory
