+++ plugins/extractors/elastic/meta.yaml
@@ -1,5 +1,5 @@
-name: "elastic"
-version: "1.0.0"
-description: "descr"
+title: "Elastic Search"
+version: "v1"
+description: "Search engine based on the Lucene library."
 sample_config: |-
   host: elastic_server
