+++ plugins/extractors/metabase/meta.yaml
@@ -1,6 +1,6 @@
-name: "metabase"
-version: "1.0.0"
-description: "descr"
+title: "Metabase"
+version: "v1"
+description: "Open source business intelligence tool."
 sample_config: |-
   host: http://localhost:3000
   user_id: meteor_tester
