+++ plugins/extractors/mongodb/meta.yaml
@@ -1,6 +1,6 @@
-name: "mongodb"
-version: "1.0.0"
-description: "descr"
+title: "MongoDB"
+version: "v1"
+description: "Document-oriented database program"
 sample_config: |-
   host: localhost:27017
   user_id: admin
