+++ plugins/extractors/mssql/meta.yaml
@@ -1,6 +1,6 @@
-name: "mssql"
-version: "1.0.0"
-description: "descr"
+title: "Microsoft SQL Server"
+version: "v1"
+description: "Relational database management system."
 sample_config: |-
   host: localhost:1433
   user_id: admin
