+++ plugins/extractors/mysql/meta.yaml
@@ -1,6 +1,6 @@
-name: "mysql"
-version: "1.0.0"
-description: "descr"
+title: "MySQL"
+version: "v1"
+description: "Open-source relational database management system"
 sample_config: |-
   host: localhost:3306
   user_id: admin
