+++ plugins/extractors/postgres/meta.yaml
@@ -1,6 +1,6 @@
-name: "postgres"
-version: "1.0.0"
-description: "descr"
+title: "PostgreSQL"
+version: "v1"
+description: "Open source object-relational database system."
 sample_config: |-
   host: localhost:5432
   user_id: admin
