+++ plugins/sinks/console/sink.go
@@ -2,6 +2,7 @@ package console
 
 import (
 	"context"
+	_ "embed"
 	"encoding/json"
 	"fmt"
 
@@ -9,6 +10,9 @@ import (
 	"github.com/odpf/meteor/registry"
 	"github.com/odpf/salt/log"
 )
+
+//go:embed meta.yaml
+var metaFile string
 
 type Sink struct {
 	logger log.Logger
@@ -18,8 +22,12 @@ func New() plugins.Syncer {
 	return new(Sink)
 }
 
-func (s *Sink) ValidateConfig(configMap map[string]interface{}) (err error) {
-	return
+func (s *Sink) Info() (plugins.Info, error) {
+	return plugins.ParseInfo(metaFile)
+}
+
+func (s *Sink) Validate(configMap map[string]interface{}) (err error) {
+	return nil
 }
 
 func (s *Sink) Sink(ctx context.Context, config map[string]interface{}, out <-chan interface{}) (err error) {
@@ -45,7 +53,7 @@ func init() {
 		return &Sink{
 			logger: plugins.GetLog(),
 		}
-	}, ""); err != nil {
+	}); err != nil {
 		panic(err)
 	}
 }
