+++ plugins/sinks/kafka/meta.yaml
@@ -1,7 +1,7 @@
-name: "kafka"
-version: "1.0.0"
-description: "descr"
+title: "kafka"
+version: "v1"
+description: "Distributed event streaming platform"
 sample_config: |-
   brokers: "localhost:9092"
-  topic: go-data
+  topic: sample-topic-name
   key_path: xxx
