+++ registry/extractors_test.go
